Make Zendesk App runnable in AWS Lamda
