+++ package.json
@@ -6,9 +6,11 @@
   "scripts": {
     "build": "tsc -p .",
     "build:watch": "tsc -w -p .",
-    "start": "nodemon dist/src/index.js",
+    "clean": "rm -rf dist node_modules",
+    "package": "rm -f bundle.zip; rm -rf dist; npm run build; cp -r node_modules dist; cd dist; zip -qr js-function * node_modules; cp ../manifest.json .; zip -r bundle.zip js-function.zip manifest.json",
+      "start": "LOCAL=true nodemon dist/index.js",
     "test": "jest --coverage",
-    "rebuild": "rm -rf node_modules; rm package-lock.json; npm i; npm run build"
+    "rebuild": "rm -rf node_modules; rm -f package-lock.json; npm i; npm run build"
   },
   "author": "",
   "license": "ISC",
@@ -18,6 +20,7 @@
     "dotenv": "8.2.0",
     "node-fetch": "2.6.1",
     "node-zendesk": "2.0.3",
+    "serverless-http": "2.7.0",
     "ts-node": "9.1.1"
   },
   "devDependencies": {
