+++ src/index.ts
@@ -1,14 +1,13 @@
 import dotenv from 'dotenv';
 dotenv.config();
 import express from 'express';
+import serverless from 'serverless-http';
 import bodyParser from 'body-parser';
-import cors from 'cors';
 
 import zdRoutes from './restapi/zd_routes';
 import mmRoutes from './restapi/mm_routes';
 
 const app = express();
-app.use(cors({origin: 'http://localhost:8065', credentials: true}));
 
 app.use(bodyParser.json());
 app.use(bodyParser.urlencoded({
@@ -18,5 +17,23 @@ app.use(bodyParser.urlencoded({
 app.use('/zendesk', zdRoutes); // Zendesk router
 app.use('/mattermost', mmRoutes); // Mattermost router
 
-const port = process.env.PORT || 4000;
-app.listen(port, () => console.log('Listening on ' + port));
+if (isRunningInHTTPMode()) {
+    // Listen to http port
+    const port = getPort();
+    app.listen(port, () => console.log('Listening on ' + port));
+} else {
+    // Export handle for aws lambda
+    module.exports.handler = serverless(app);
+}
+
+export function isRunningInHTTPMode(): boolean {
+    return process.env.LOCAL === 'true'
+}
+
+function getPort(): number {
+    return Number(process.env.PORT) || 4000;
+}
+
+export function getHTTPPath(): string {
+    return 'http://localhost:' + getPort()
+}
