+++ src/restapi/fManifest.ts
@@ -1,6 +1,6 @@
 import {Request, Response} from 'express';
 
-import {getManifest} from '../../manifest';
+import {getManifest} from '../manifest';
 
 export function fManifest(_: Request, res: Response): void {
     res.json(getManifest());
